requirements.txt
flask>=3.0,<4
python-dotenv>=1.0,<2
supabase>=2.10,<3
httpx>=0.27,<0.29
